requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
groq==0.4.2
langgraph==0.0.18
langchain==0.1.3
langchain-groq==0.0.11
httpx==0.25.1
PyPDF2==3.17.1
langsmith==0.0.78
python-multipart==0.0.6
aiohttp==3.9.1
